Defined test helpers
